Advanced Spreadsheets Project
In this project the task was to act as a consultant for a client to analyze the vacation rental market in the Manhattan borough of New York City and provide useful insights as to where is the best place to invest in a property.

I started by exploring, cleaning, and filtering the data. Then Analyzing which neighborhoods are the most attractive, which size properties (how many bedrooms) were most popular,  calculating occupancy, and estimating the revenue for investment properties, while documenting the changes I made, putting together a table of contents, and writing an executive summary.