What is causing returns?
The project task was to determine what was causing the high amount of returns at a superstore.

I started by familiarizing myself with the given data. Then I built several worksheets to analyze different views on return rates, built a tableau dashboard to tell the story of the data, then recorded a presentation of the visualizations and my findings.