Zomato Sales Analysis
The task for this project was to analyze the business performance of a company named Zomato.

For this project I received raw and unorganized data. I started by familiarizing myself with the data and organizing and cleaning. I wrote a well organized report and built a tableau dashboard to tell the story of the data. The link to the Tableau dashboard can be found at the bottom of the report.